requirements.txt
requests
beautifulsoup4
lxml
python-whois
Flask 